+++ src/components/Layout.styl
@@ -2,8 +2,6 @@
   margin-top: 20px;
 }
 
-
-
 .accordion.ui.styled.faq-accordion {
   .title {
     color: rgba(0,0,0,.87);
