+++ src/pages/404.js
@@ -1,8 +1,8 @@
 import React from 'react';
 
 import { Link } from 'gatsby';
-import Layout from '../components/layout';
-import SEO from '../components/seo';
+import Layout from '../components/Layout';
+import SEO from '../components/SEO';
 
 const NotFoundPage = () => (
   <Layout>
